# openapi.yaml
openapi: 3.1.0
info:
  title: SISR
  description: ''
  license:
    name: ''
  version: 0.0.1
paths:
  /api/v1/config:
    get:
      tags:
      - config
      summary: Get Config File
      description: |-
        Returns the contents of the config file at the default config path.
        Fields not present in the file are omitted.
      operationId: get_config_file
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigFileData'
        '404':
          description: No config file found at default path
    patch:
      tags:
      - config
      summary: Update Config File
      description: |-
        Merges the provided fields into the config file at the default config path.
        Only fields present in the request body are written; all fields are optional.
      operationId: update_config_file
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigFileData'
        required: true
      responses:
        '200':
          description: OK
...
  /api/v1/steam_shortcuts:
    get:
      tags:
      - steam
      summary: List Steam Shortcuts
      description: Lists the active Steam user's non-Steam shortcuts so SISR can switch Steam Input profiles.
      operationId: get_steam_shortcuts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SteamShortcutsResponse'
        '409':
          description: Steam shortcut data unavailable
  /api/v1/steam_shortcuts/launch:
    post:
      tags:
      - steam
      summary: Launch Steam Shortcut
      description: Persists the selected shortcut app id, asks Steam to launch it, then shuts this SISR instance down.
      operationId: launch_steam_shortcut
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectSteamShortcutRequest'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Unknown shortcut app id
        '500':
          description: Failed to launch Steam shortcut
        '409':
          description: Steam shortcut data unavailable
...
  /api/v1/shutdown:
    post:
      tags:
      - ui
      summary: Shutdown Application
      description: Shuts down the application gracefully
      operationId: shutdown
      responses:
        '200':
          description: ''
        '500':
          description: Unknown error
...
components:
  schemas:
...
    ConfigFileControllerEmulation:
      type: object
      properties:
        allow_desktop_config:
          type:
          - boolean
          - 'null'
          default: null
        default_controller_type:
          type:
          - string
          - 'null'
          default: null
        gyro_passthrough:
          type:
          - boolean
          - 'null'
          default: null
        steam_input_profile_app_id:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
          default: null
        require_controllers_connected_before_launch:
          type:
          - boolean
          - 'null'
          default: null
    ConfigFileData:
      type: object
      properties:
        controller_emulation:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ConfigFileControllerEmulation'
          default: null
        kbm_emulation:
          type:
          - boolean
          - 'null'
          default: null
        log:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ConfigFileLogOpts'
          default: null
        port:
          type:
          - integer
          - 'null'
          format: int32
          default: null
          minimum: 0
        steam:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ConfigFileSteamOpts'
          default: null
        tray:
          type:
          - boolean
          - 'null'
          default: null
        update_notify:
          type:
          - string
          - 'null'
          default: null
        viiper_address:
          type:
          - string
          - 'null'
          default: null
        viiper_password:
          type:
          - string
          - 'null'
          default: null
        window:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ConfigFileWindowOpts'
          default: null
    ConfigFileLogOpts:
      type: object
      properties:
        level:
          type:
          - string
          - 'null'
          default: null
    ConfigFileSteamOpts:
      type: object
      properties:
        no_steam:
          type:
          - boolean
          - 'null'
          default: null
        steam_path:
          type:
          - string
          - 'null'
          default: null
    ConfigFileWindowOpts:
      type: object
      properties:
        continuous_draw:
          type:
          - boolean
          - 'null'
          default: null
        create:
          type:
          - boolean
          - 'null'
          default: null
        fullscreen:
          type:
          - boolean
          - 'null'
          default: null
...
    SelectSteamShortcutRequest:
      type: object
      required:
      - app_id
      properties:
        app_id:
          type: integer
          format: int32
          minimum: 0
...
    SteamShortcutProfile:
      type: object
      required:
      - app_id
      - game_id
      - name
      - exe
      - selected
      properties:
        app_id:
          type: integer
          format: int32
          minimum: 0
        game_id:
          type: integer
          format: int64
          minimum: 0
        exe:
          type: string
        launch_options:
          type:
          - string
          - 'null'
        name:
          type: string
        selected:
          type: boolean
        start_dir:
          type:
          - string
          - 'null'
    SteamShortcutsResponse:
      type: object
      required:
      - shortcuts_path
      - shortcuts
      properties:
        selected_app_id:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        shortcuts:
          type: array
          items:
            $ref: '#/components/schemas/SteamShortcutProfile'
        shortcuts_path:
          type: string
...